feat: MS-DOS backend (net_dos.c + fsplat_dos.c, DJGPP) - step 2 of the port
Add the DOS backends behind the net.h/fsplat.h shims, completing the
three-platform matrix (Linux / Win9x / DOS). The server builds as a 32-bit
protected-mode DJGPP program (coff-go32-exe; CWSDPMI.EXE as the runtime DPMI
host on pure DOS) - required because the fsdb[65536] cache alone exceeds the
real-mode 640K limit, and trivial in DJGPP's flat address space.

- net_dos.c: hand-rolled Crynwr/FTP packet-driver client ("Option A", no
  Watt-32/NASM). Scans INT 60h-80h for the "PKT DRVR" signature, validates via
  driver_info (AH=01h), registers the 0xEDF5 ethertype with access_type
  (AH=02h), sends with send_pkt (AH=04h), releases with release_type (AH=03h).
  The receiver is a DPMI real-mode RETF callback (the driver FAR-CALLs it):
  AX=0 returns a conventional-memory bounce buffer, AX=1 copies the frame into
  a single-producer/single-consumer ring that net_recv() drains. Three
  separate DOS buffers (RX bounce / TX / ethertype template) so a receive
  burst can never clobber a frame mid-send or the driver's filter template.
  Whole image locked via _CRT0_FLAG_LOCK_MEMORY (+ span locks) so nothing the
  interrupt-time callback touches can be paged out.

- fsplat_dos.c: DJGPP's POSIX-ish libc covers most of the interface
  (opendir/readdir with d_type, stat, statvfs, truncate, mkdir/rmdir/chdir/
  unlink/rename); DOS attributes via _chmod (INT 21h AX=4300h/4301h),
  _fixpath + stat for plat_fullpath, plat_isfat trivially 0 (DOS is FAT).

- Makefile.dos (i586-pc-msdosdjgpp-gcc, -march=i386 for pre-Pentium hardware)
  and Dockerfile.djgpp (reproducible cross-toolchain: andrewwutw build-djgpp
  v3.4 prebuilt, GCC 12.2.0). DJGPP getopt/snprintf are native - no
  compat_getopt.c and compat_stdio.h stays a no-op here.

The implementation was adversarially reviewed against the Packet Driver
Specification, RBIL and the DJGPP libc reference; confirmed findings fixed:
  * BIOS tick counter (0040:006C) resets to 0 at midnight (not a power-of-2
    wrap): absolute-deadline timeouts could hang net_recv forever on an idle
    link. Timing is now elapsed-based with explicit reset detection
    (BIOS_TICKS_PER_DAY).
  * ms_to_ticks rounded sub-55ms timeouts down to 0 ticks (instant return,
    no yield): now rounds up with a 1-tick floor.
  * get_address wrote the MAC into the live RX bounce buffer while the
    receiver was already installed - an inbound frame could clobber it; the
    MAC now lands in tx_seg, which the receiver never touches.
  * pkt_handle==0 doubled as the not-open sentinel while 0 can be a valid
    driver handle; a separate handle_open flag now gates release_type.

Also: two DBG format strings use %ld+cast (DJGPP int32_t is long). Linux
build + full regression suite unchanged and green; Win9x build stays
wide-import-free.

Runtime validation on a real packet driver (DOSBox-X/86Box NE2000, then
vintage hardware) is intentionally deferred - this commit is the
compile/link-verified, spec-reviewed implementation.

+++ ethersrv.c
@@ -464,8 +464,8 @@ static int process(struct struct_answcache *answer, unsigned char *reqbuff,
     offset = le32toh(((uint32_t *)reqbuff)[0]);
     fileid = le16toh(wreqbuff[2]);
     len = le16toh(wreqbuff[3]);
-    DBG("Asking for %u bytes of the file #%u, starting offset %u\n", len,
-        fileid, offset);
+    DBG("Asking for %u bytes of the file #%u, starting offset %lu\n", len,
+        fileid, (unsigned long)offset);
     readlen = readfile(answ, fileid, offset, len);
     if (readlen < 0) {
       fprintf(stderr, "ERROR: invalid handle\n");
@@ -483,8 +483,8 @@ static int process(struct struct_answcache *answer, unsigned char *reqbuff,
       long writelen;
       offset = le32toh(((uint32_t *)reqbuff)[0]);
       fileid = le16toh(wreqbuff[2]);
-      DBG("Writing %u bytes into file #%u, starting offset %u\n",
-          reqbufflen - 6, fileid, offset);
+      DBG("Writing %u bytes into file #%u, starting offset %lu\n",
+          reqbufflen - 6, fileid, (unsigned long)offset);
       writelen = writefile(reqbuff + 6, fileid, offset, reqbufflen - 6);
       if (writelen < 0) {
         *ax = 5; /* "access denied" */
@@ -989,7 +989,7 @@ static int process(struct struct_answcache *answer, unsigned char *reqbuff,
     int32_t offs = le32toh(((uint32_t *)reqbuff)[0]);
     long fsize;
     unsigned short fss = le16toh(((unsigned short *)reqbuff)[2]);
-    DBG("SKFMEND on file #%u at offset %d\n", fss, offs);
+    DBG("SKFMEND on file #%u at offset %ld\n", fss, (long)offs);
     /* if arg is positive, zero it out */
     if (offs > 0)
       offs = 0;
@@ -1003,7 +1003,7 @@ static int process(struct struct_answcache *answer, unsigned char *reqbuff,
       offs += fsize;
       if (offs < 0)
         offs = 0;
-      DBG("new offset: %d\n", offs);
+      DBG("new offset: %ld\n", (long)offs);
       ((uint32_t *)answ)[0] = htole32(offs);
       reslen = 4;
     }
